+++ src/store/queries/selectors.ts
@@ -1,9 +1,42 @@
+import { createSelector } from 'reselect'
 import { State } from '../reducer';
+import decodeQueryCacheKey from '../../utils/decodeQueryCacheKey';
 
-export const selectQuery = (state: State, key: string) => {
-  return state.queries[key];
-};
+export const selectQueries = (state: State) => state.queries;
 
-export const selectMetaFromQuery = (state: State, key: string) => {
-  return selectQuery(state, key).meta;
-};
+export const selectQueryByKey = createSelector(
+  selectQueries,
+  (_, key) => key,
+  (queries, key) => queries[key]
+);
+
+export const selectIdsFromQuery = createSelector(selectQueryByKey, (query) => query && query.ids);
+
+export const selectDomainNameFromQuery = createSelector(
+  (_, key) => key,
+  (key) => decodeQueryCacheKey(key).domain,
+);
+
+export const selectDataFromQuery = createSelector(selectQueryByKey, (query) => query && query.data);
+
+// Factories
+
+export const selectLoadingFromQueryFactory = () => createSelector(
+  selectQueryByKey,
+  (query) => query ? query.loading : true
+);
+
+export const selectMetaFromQueryFactory = () => createSelector(
+  selectQueryByKey,
+  (query) => query && query.meta
+);
+
+export const selectLoadingFromQueryWithUpdateQueriesFactory = () => createSelector(
+  (state: State) => state,
+  (_, key) => key,
+  (_, __, updateQueryKeys) => updateQueryKeys,
+  (state, key: string, updateQueryKeys: string[]) => {
+    const selectLoadingFromQuery = selectLoadingFromQueryFactory();
+    return [key, ...updateQueryKeys].some(nextKey => selectLoadingFromQuery(state, nextKey));
+  }
+)