+++ src/utils/useUpdatableState/index.ts
@@ -1,2 +1,0 @@
-import useUpdatableState from './useUpdatableState';
-export default useUpdatableState;